requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
openai
PyPDF2
gradio